ultima carga

+++ js/carrito.js
@@ -28,6 +28,11 @@ const numerito = document.querySelector("#numerito");
 
 
 
+
+
+
+
+
 function cargarProductosCarrito(){
 
     if (productosEnCarrito && productosEnCarrito.length > 0) {
@@ -149,6 +154,26 @@ function eliminarDelCarrito(e){
 
     localStorage.setItem("productos-en-carrito",JSON.stringify(productosEnCarrito));
 
+
+    Toastify({
+        text: "¡Listo! Eliminaste el producto!",
+        duration: 2000,
+        gravity: "bottom", // `top` or `bottom`
+        position: "center", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background: "linear-gradient(to right, #F35200, #F35200)",
+          borderRadius: "0.5rem",
+        },
+        offset: {
+            x: '2rem', // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+            y: '3rem' // vertical axis - can be a number or a string indicating unity. eg: '2em'
+          },
+        onClick: function(){} // Callback after click
+      }).showToast();
+
+
+
 }
 
 
@@ -199,29 +224,30 @@ function restarDelCarrito(e){
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 botonVaciar.addEventListener("click", vaciarCarrito);
 
 function vaciarCarrito (){
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-    cargarProductosCarrito();
     
+    
+    //Libreria sweetalert
 
+    Swal.fire({
+        title: 'Vas a vaciar tu carrito',
+        html:'¿Estas seguto?',
+        showCloseButton: true,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText:'Si',
+        confirmButtonColor: '#F35200',
+        cancelButtonColor: '#a8a6a6',
+        cancelButtonText:'No'        
+      }).then((result) => {
+        if (result.isConfirmed){
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+            cargarProductosCarrito();
+        }
+      })
 
 }
 
@@ -245,3 +271,53 @@ function comprarCarrito() {
     contenedorCarritoComprado.classList.remove("disabled");
 
 }
+
+
+
+
+const monedaEl_one = document.getElementById('moneda-uno');
+const monedaEl_two = document.getElementById('moneda-dos');
+const cantidadEl_one = document.getElementById('cantidad-uno');
+const cantidadEl_two = document.getElementById('cantidad-dos');
+
+const cambioEl = document.getElementById('cambio');
+
+const tazaEl = document.getElementById('taza');
+
+
+
+
+
+
+async function calculate(){
+    const moneda_one = monedaEl_one.value;
+    const moneda_two = monedaEl_two.value;
+
+
+
+    const cambioapi = await fetch(`https://api.exchangerate-api.com/v4/latest/${moneda_one}`);
+    const data = await cambioapi.json();
+
+       const taza = data.rates[moneda_two];
+       cambioEl.innerText = `1 ${moneda_one} = ${taza} ${moneda_two}`;
+       cantidadEl_two.value = (cantidadEl_one.value * taza).toFixed(2);
+   
+    
+}
+
+monedaEl_one.addEventListener('change', calculate);
+cantidadEl_one.addEventListener('input', calculate);
+monedaEl_two.addEventListener('change', calculate);
+cantidadEl_two.addEventListener('input', calculate);
+
+taza.addEventListener('click', () =>{
+    const temp = monedaEl_one.value;
+    monedaEl_one.value = monedaEl_two.value;
+    monedaEl_two.value = temp;
+    calculate();
+} );
+
+
+calculate();
+
+
